requirements.txt
click==6.7
Flask==1.0.2
Flask-Script==2.0.6
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Pillow==5.2.0
pkg-resources==0.0.0
pyocr==0.5.2
six==1.11.0
Werkzeug==0.14.1
